Use pure-python types instead of numpy
This change updates method signatures and methods to use pure python types instead of numpy types. It also removes numpy as a dependency, as it is no longer used.

+++ tests/test_cpu_bus.py
@@ -1,4 +1,3 @@
-import numpy as np
 import pytest
 
 from purenes.cpu import CPUBus
@@ -16,16 +15,16 @@ class TestCpuBus(object):
         yield cpu_bus
 
     def test_read_from_ram(self, test_object):
-        address_range = [np.uint16(x) for x in range(0x0000, 0x2000)]
+        address_range = [x for x in range(0x0000, 0x2000)]
 
         for address in address_range:
-            data: np.uint8 = test_object.read(address)
+            data: int = test_object.read(address)
 
             assert(data == 0x00)
 
     def test_write_to_ram(self, test_object):
-        address_range = [np.uint16(x) for x in range(0x0000, 0x2000)]
-        data = np.uint8(0x01)
+        address_range = [x for x in range(0x0000, 0x2000)]
+        data = 0x01
 
         for address in address_range:
             test_object.write(address, data)
@@ -33,7 +32,7 @@ class TestCpuBus(object):
             assert(test_object.read(address) == data)
 
     def test_read_invalid_address_raises_exception(self, test_object):
-        invalid_address = np.uint32(0x10000)
+        invalid_address = 0x10000
 
         with pytest.raises(Exception) as exception:
             test_object.read(invalid_address)
@@ -41,9 +40,9 @@ class TestCpuBus(object):
         assert(str(exception.value) == self.INVALID_ADDRESS_EXCEPTION_MESSAGE)
 
     def test_write_invalid_address_raises_exception(self, test_object):
-        invalid_address = np.uint32(0x10000)
+        invalid_address = 0x10000
 
         with pytest.raises(Exception) as exception:
-            test_object.write(invalid_address, np.uint8(0x01))
+            test_object.write(invalid_address, 0x01)
 
         assert(str(exception.value) == self.INVALID_ADDRESS_EXCEPTION_MESSAGE)
